tests/wpt/web-platform-tests/feature-policy/feature-policy-for-sandbox/sandbox-policies-in-allow-attribute.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="iframe-embedder"></div>
<script src="./resources/helper.js"></script>
<script>
'use strict';

const iframe_src = "/feature-policy/feature-policy-for-sandbox/resources/window_opener.html";

promise_test( async () => {
  const iframe = await add_iframe({
    src: iframe_src,
    allow: sandbox_features.map(feature => `${feature} *`).join(", "),
    sandbox: ""
  });

  iframe.contentWindow.postMessage({type: "features", features: sandbox_features}, "*");
  const iframe_states = await feature_update();
  for (const [feature, state] of iframe_states)
    assert_true(state, `'${feature}' should not be disabled in <iframe>.'`);

  iframe.parentElement.removeChild(iframe);
}, "Verify that when a sandbox related feature is enabled in 'allow' then " +
   " the feature will be enabled regardless of sandbox attribute's value.");

promise_test( async () => {
  const non_script_sandbox_features =
    sandbox_features.filter(feature => feature !== "scripts");
  const sandbox_flags =
    sandbox_features.map(feature => `allow-${feature}`).join(" ");
  const allow_attribute =
    non_script_sandbox_features.map(feature => `${feature} 'none'`).join(", ");

  const iframe = await add_iframe(
      {src: iframe_src, allow: allow_attribute, sandbox: sandbox_flags});

  iframe.contentWindow.postMessage(
    {type: "features", features: non_script_sandbox_features}, "*");
  const iframe_states = await feature_update();
  for (const [feature, state] of iframe_states)
    assert_false(state, `'${feature}' should be disabled in <iframe>.'`);

  iframe.parentElement.removeChild(iframe);
}, "Verify that when a sandbox related feature (except scripts) is disabled in 'allow' then " +
   " the feature will be disabled regardless of sandbox attribute's value.");

promise_test( async () => {
  const iframe = await add_iframe(
    {src: iframe_src, allow: "scripts 'none'", sandbox: "allow-scripts"});

  iframe.contentWindow.postMessage(
    {type: "features", features: ['scripts']}, "*");
  const iframe_states = await feature_update(10);

  assert_equals(iframe_states.length, 0);
  iframe.parentElement.removeChild(iframe);
}, "Verify that when scripts is disabled in 'allow' then " +
  " the feature will be disabled regardless of sandbox attribute's value.");
</script>
</body>
